refactor: move parse util into package file
