--- src/utils/parseRepositoryUrl.ts
@@ -1,9 +1,0 @@
-export function parseRepositoryUrl(url: string) {
-  try {
-    const { protocol, href } = new URL(url.replace(/.git$/, ''));
-
-    return href.replace(protocol, 'https:');
-  } catch (error) {}
-
-  return null;
-}
